# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Usuarios y Pagos
  version: 2.0.0
  description: API RESTful para la gestión de usuarios con roles, permisos, autenticación y sistema de pagos PayPal
  contact:
    name: Soporte API
    email: soporte@empresa.com
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
...
paths:
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Obtener todos los usuarios
      description: Retorna una lista de todos los usuarios registrados
      responses:
        '200':
          description: Lista de usuarios obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Usuario'
                  total:
                    type: integer
                    example: 3
        '401':
          description: No autorizado

    post:
      tags: [Usuarios]
      summary: Crear un nuevo usuario
      description: Crea un nuevo usuario en el sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioInput'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Error de validación
        '401':
          description: No autorizado
...
  /usuarios/roles:
    get:
      tags: [Usuarios]
      summary: Obtener roles disponibles
      description: Retorna la lista de roles disponibles para los usuarios
      responses:
        '200':
          description: Lista de roles obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                    example: [administrador, editor, usuario]
        '401':
          description: No autorizado
  /usuarios/estadisticas:
    get:
      tags: [Usuarios]
      summary: Obtener estadísticas de usuarios
      description: Retorna estadísticas generales sobre los usuarios
      responses:
        '200':
          description: Estadísticas obtenidas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 15
                      activos:
                        type: integer
                        example: 12
                      inactivos:
                        type: integer
                        example: 3
                      por_rol:
                        type: object
                        example:
                          administrador: 2
                          editor: 5
                          usuario: 8
        '401':
          description: No autorizado
...
  /usuarios/preferencias:
    get:
      tags: [Preferencias]
      summary: Obtener preferencias del usuario
      description: Retorna las preferencias del usuario autenticado
      responses:
        '200':
          description: Preferencias obtenidas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PreferenciasUsuario'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Preferencias]
      summary: Actualizar todas las preferencias del usuario
      description: Actualiza completamente las preferencias del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenciasInput'
      responses:
        '200':
          description: Preferencias actualizadas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PreferenciasUsuario'
                  message:
                    type: string
                    example: Preferencias actualizadas exitosamente
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ValidationError'
        '401':
          description: No autorizado
...
components:
  schemas:
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          format: email
          example: juan.perez@email.com
        rol:
          type: string
          example: administrador
        fecha_creacion:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00.000Z
        current_latitude:
          type: number
          format: float
          example: 19.432608
        current_longitude:
          type: number
          format: float
          example: -99.133209
    UsuarioInput:
      type: object
      required:
        - nombre
        - email
        - password
        - rol
      properties:
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          format: email
          example: juan.perez@email.com
        password:
          type: string
          example: Password123
        rol:
          type: string
          example: administrador
        activo:
          type: boolean
          example: true

    # Schemas de Pagos
...
    PreferenciasUsuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 14
        theme:
          type: string
          enum: [light, dark, auto]
          example: "dark"
        language:
          type: string
          example: "es"
        notifications:
          type: object
          properties:
            email:
              type: boolean
              example: true
            push:
              type: boolean
              example: false
            sms:
              type: boolean
              example: false
        privacy:
          type: object
          properties:
            showLocation:
              type: boolean
              example: false
            showOnlineStatus:
              type: boolean
              example: true
            allowContactByEmail:
              type: boolean
              example: true
        dashboard:
          type: object
          properties:
            layout:
              type: string
              enum: [grid, list, cards]
              example: "grid"
            itemsPerPage:
              type: integer
              minimum: 5
              maximum: 100
              example: 20
        created_at:
          type: string
          format: date-time
          example: "2025-09-24T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-09-24T10:35:00Z"
    PreferenciasInput:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, auto]
          example: "dark"
        language:
          type: string
          example: "es"
        notifications:
          type: object
          properties:
            email:
              type: boolean
              example: true
            push:
              type: boolean
              example: false
            sms:
              type: boolean
              example: false
        privacy:
          type: object
          properties:
            showLocation:
              type: boolean
              example: false
            showOnlineStatus:
              type: boolean
              example: true
            allowContactByEmail:
              type: boolean
              example: true
        dashboard:
          type: object
          properties:
            layout:
              type: string
              enum: [grid, list, cards]
              example: "grid"
            itemsPerPage:
              type: integer
              minimum: 5
              maximum: 100
              example: 20
...
  responses:
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: No autorizado
              error:
                type: string
                example: Token inválido o expirado
...
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Error de validación
              errors:
                type: array
                items:
                  type: string
                example: 
                  - El email es obligatorio
                  - La contraseña debe tener al menos 6 caracteres
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
security:
  - BearerAuth: []